Upgrade Selenium.WebDriver to 4.6.0
