+++ src/ServerHelper.cs
@@ -4,7 +4,6 @@ using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Net.NetworkInformation;
-using CSharpGalenWrapper.Driver;
 using CSharpGalenWrapper.Layout;
 using Newtonsoft.Json;
 using OpenQA.Selenium;
